pyproject.toml
[project]
name = "kdx-pi-cam"
version = "1.0.0"
description = "Aplicación Python asíncrona que procesa video de una cámara RTSP bajo demanda y notifica eventos a través de Telegram"
authors = [{name = "KDX Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
kdx-pi-cam = "main:main"